# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SIH SaaS API — Contrat HTTP complet de la Phase 0 (SaaS Core) et des trois premiers increments de Phase 1
  version: 0.6.0
  description: >
    Premier fichier OpenAPI de ce depot (Phase 0, etape 5/13 : confirmation de paiement
    d'abonnement SaaS serveur-a-serveur, O-25.5). Etendu a l'etape 11/13 (ADR-0009) avec le
    PREMIER endpoint HTTP AUTHENTIFIE du depot : consultation du journal d'audit plateforme
    (`GET /api/v1/audit-entries`). Etendu a l'etape 12/13 item 11 (gap transverse, pas specifique
    a l'isolation/securite) pour couvrir les 10 routes reellement montees dans `createApp()`
    (`apps/api/src/server.ts`) : inscription/connexion/MFA (ADR-0010), recuperation break-glass
    `SUPER_ADMIN` (ADR-0005 Amendement 1) et `/health`. Etendu au premier increment vertical de
    Phase 1 (permission `tenant-config:administer`, decision de gouvernance 2026-09-28 sur le
    passage Phase 0 -> Phase 1) avec `GET`/`PATCH /api/v1/facility` : consultation et renommage de
    l'identite de l'etablissement du tenant courant (`HealthFacility`), PREMIER endpoint HTTP du
    module `tenant`. Etendu au deuxieme increment vertical de Phase 1 (permission
    `membership:administer`, meme decision de gouvernance) avec `GET /api/v1/memberships` :
    consultation paginee des memberships de l'etablissement du tenant courant, PREMIER endpoint
    HTTP de ce type pour le module `identity`. Document PUREMENT DESCRIPTIF : aucune
...
servers:
...
paths:
...
  /api/v1/platform/tenant-admin-recovery-requests:
    post:
      operationId: requestTenantAdminRecovery
      tags: [PlatformTenantAdminRecovery]
      summary: Demander la récupération de l'accès administrateur d'un tenant
      description: >
        Route plateforme authentifiée : session `PLATFORM` (SUPER_ADMIN) avec MFA satisfaite.
        Cette action ouvre une demande sans accorder de membership. L'approbation doit venir d'un
        autre SUPER_ADMIN. Avant approbation, deux SUPER_ADMIN supplémentaires, distincts du
        demandeur et de l'approbateur, doivent avoir effectué la vérification
        d'identité hors bande prévue par la gouvernance. La méthode concrète reste à définir ;
        l'API ne collecte ni ne vérifie ces attestations.
      security:
        - bearerSessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, beneficiaryUserId, reason]
              additionalProperties: false
              properties:
                tenantId:
                  type: string
                  format: uuid
                beneficiaryUserId:
                  type: string
                  format: uuid
                reason:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        '201':
          description: Demande créée, en attente de l'approbation d'un second SUPER_ADMIN.
          content:
            application/json:
              schema:
                type: object
                required: [requestId, status]
                properties:
                  requestId:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [PENDING]
        '400':
          description: Corps invalide ou raison vide après normalisation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: invalid_request
        '401':
          description: Aucune session valide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: unauthenticated
        '403':
          description: Session valide, mais pas PLATFORM avec MFA satisfaite.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: forbidden
        '404':
          description: Tenant ou compte bénéficiaire introuvable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: not_found
        '500':
          description: Défaillance technique.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: internal_error
  /api/v1/platform/tenant-admin-recovery-requests/{requestId}/approval:
    post:
      operationId: approveTenantAdminRecovery
      tags: [PlatformTenantAdminRecovery]
      summary: Approuver une demande de récupération d'accès administrateur
      description: >
        Session PLATFORM avec MFA satisfaite, distincte du demandeur. La mutation et les deux
        écritures d'audit sont atomiques. L'opération ne réussit que si aucun membership actif du
        tenant ne porte `membership:administer`, y compris via un rôle personnalisé. Les deux
        SUPER_ADMIN chargés de la vérification d'identité hors bande doivent être distincts du
        demandeur et de l'approbateur ; cette condition humaine n'est pas validée par l'API et sa
        procédure concrète reste à définir.
      security:
        - bearerSessionAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              description: Doit être absent ou un objet vide.
      responses:
        '200':
          description: Demande approuvée et accès administrateur attribué.
          content:
            application/json:
              schema:
                type: object
                required: [status, membershipId]
                properties:
                  status:
                    type: string
                    enum: [approved]
                  membershipId:
                    type: string
                    format: uuid
        '400':
          description: Identifiant invalide ou corps non vide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: invalid_request
        '401':
          description: Aucune session valide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: unauthenticated
        '403':
          description: Session valide, mais pas PLATFORM avec MFA satisfaite.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: forbidden
        '404':
          description: Demande, tenant ou compte bénéficiaire introuvable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: not_found
        '409':
          description: Demande déjà traitée, demandeur approbateur, ou tenant ayant déjà un administrateur.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: conflict
        '500':
          description: Défaillance technique.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleError'
              example:
                error: internal_error
...
components:
  securitySchemes:
    bearerSessionAuth:
      type: http
      scheme: bearer
      description: >
        `sessionId` opaque (jamais un JWT auto-porteur) — ADR-0009 §8.3. Choix delibere et
        minimaliste, explicitement revisable a l'arrivee d'`apps/web` (transport definitif du
        `sessionId`, residu 5 de l'ADR).
  schemas:
...
    SimpleError:
      type: object
      description: >
        Format d'erreur produit par TOUTES les routes de ce contrat (`createErrorHandler`,
        `src/server.ts`, et chaque controleur individuellement) — PAS `ProblemDetails`
        (RFC 9457), jamais de detail interne expose. Ne pas confondre avec `ProblemDetails`
        ci-dessus, reserve aux futurs endpoints CRUD. L'enumeration ci-dessous est l'UNION de
        tous les codes reellement retournes par une route de ce fichier — une route donnee n'en
        retourne jamais qu'un sous-ensemble (voir ses reponses documentees).
      required: [error]
      properties:
        error:
          type: string
          enum:
            - invalid_request
            - unauthenticated
            - invalid_credentials
            - forbidden
            - mfa_required
            - mfa_enrollment_required
            - mfa_enrollment_already_active
            - email_already_registered
            - not_found
            - conflict
            - too_many_requests
            - internal_error
...